+++ front-end/src/dashboard/DashboardReservationList.js
@@ -5,56 +5,9 @@ import ErrorAlert from "../layout/ErrorAlert";
 import DashboardReservationItem from "./DashboardReservationItem";
 import { today } from "../utils/date-time";
 
-function DashboardReservationList({date, setReservationsError, setReservations, reservations, reservationsError}) {
-    const history = useHistory()
-
-    const location = useLocation()
-    const queryParams = new URLSearchParams(location.search)
-    const reservationDate = queryParams.get('date')
-
-    if(reservationDate) {
-        date = reservationDate
-    }
-  
-    useEffect(loadDashboard, [date]);
-  
-    function loadDashboard() {
-      const abortController = new AbortController();
-      setReservationsError(null);
-      listReservations({ date }, abortController.signal)
-        .then(setReservations)
-        .catch(setReservationsError);
-      return () => abortController.abort();
-    }
-  
-    function previousHandler() {
-      const currentDate = new Date(date)
-      currentDate.setDate(currentDate.getDate() - 1)
-      const previousDate = currentDate.toISOString().slice(0,10)
-      history.push(`/dashboard/?date=${previousDate}`)
-    }
-
-    function forwardHandler() {
-      const currentDate = new Date(date)
-      currentDate.setDate(currentDate.getDate() + 1)
-      const nextDate = currentDate.toISOString().slice(0,10)
-      history.push(`/dashboard/?date=${nextDate}`)
-    }
-
-    function todayHandler() {
-      const currentDate = new Date(today())
-      history.push(`/dashboard/?date=${currentDate.toISOString().slice(0,10)}`)
-    }
-    
+function DashboardReservationList({reservations, reservationsError, statusToDisplay}) {
+   
     return <div>
-        <div className="d-md-flex mb-3">
-          <h4 className="mb-0">Reservations for {date}</h4>
-        </div>
-        <div>
-          <button onClick={previousHandler}>Previous</button>
-          <button onClick={todayHandler}>Today</button>
-          <button onClick={forwardHandler}>Next</button>
-        </div>
         <table>
           <tr>
             <th>First Name</th>
@@ -66,7 +19,7 @@ function DashboardReservationList({date, setReservationsError, setReservations, 
             <th>Status</th>
             <th>Seat Table</th>
           </tr>
-          {reservations.map(reservation=><DashboardReservationItem reservation={reservation} />)}
+          {reservations.map(reservation=><DashboardReservationItem reservation={reservation} statusToDisplay={statusToDisplay}/>)}
         </table>
         <ErrorAlert error={reservationsError} />
     </div>
